@@ -1,39 +1,23 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="References\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
-		<PackageReference Include="Lib.Harmony" Version="2.2.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="CobaltCore">
-      <HintPath>References\CobaltCore.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="CobaltCoreModding.Definitions">
-      <HintPath>References\CobaltCoreModding.Definitions.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="MonoGame.Framework">
-      <HintPath>References\MonoGame.Framework.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="sprites\*">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
+﻿<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<RootNamespace>clay.PhilipTheMechanic</RootNamespace>
+		<Version>1.0.0</Version>
+		<IsNickelMod>True</IsNickelMod>
+		<IncludedModProjectPaths>i18n;assets</IncludedModProjectPaths>
+		<LangVersion>12.0</LangVersion>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>disable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<WarningsAsErrors>Nullable</WarningsAsErrors>
+		<CheckEolTargetFramework>false</CheckEolTargetFramework>
+		<ModZipPath>$(MSBuildProjectDirectory)\.release\$(MSBuildProjectName)-$(Version).zip</ModZipPath>
+		<EnableDllExtract>false</EnableDllExtract>
+		<GameDllPath>./CobaltCore.dll</GameDllPath>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Nickel.ModBuildConfig" Version="0.4.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<Project Include="Configuration.props.user" Condition="Exists('Configuration.props.user')" />
+	</ItemGroup>
 </Project>
